playground-internal/statefun-playground-producer/requirements.txt
kafka-python
jsonpath-ng
